PP-6029 Ignore kind-of vulnerability
Ignore vulnerability in `kind-of` that snyk is complaining about. Reason:
```
reason: this is a low severity exploit that allows crafted user input to manipulate type
detection. kind-of is used at very low level by nunjucks. We validate user input using Joi,
and generally only access specified fields in njk templates. In any case exploiting this
vulnerability in this context does not seem to carry any threat.
```
Ignoring for a year as this seems a low risk.