BAU: remove oauth requirement from live dashboard route
The live payments dashboard is currently behind the secure middleware
which requires the session to have validated with GitHub OAuth before
serving the page. This requires the display machines to have access
to the other functionality provided in Toolbox. As this is served as in
internal service requiring strict internal IP access we can remove this
and allow the displays in the office to access this resource without an
OAuth session.