PP-6219 Run ECS task after file uploaded
Run the ECS task to stream the CSV to the events SQS, giving the key of
the file uploaded to S3 to process.

Report back the ARN of the task so the user can check the progress in
splunk.

The task definition is provided using the
AWS_ECS_UPDATE_TRANSACTIONS_TASK_DEFINITION environment variable.