PP-6389: fix incorrect Sentry import
Sentry is a Typescript library that doesn't have a default export, this
means that the wildcard syntax should be used to collect all of the
modules individual exports. Without this Sentry has been failing to
intialise correctly in some cases and potentially stopping the entire
winston logging pipeline out of the app.