BAU: redact result from products payment links list
The products backend service only exposes an internal `v1/api` route
with all of the information used by other services. Currently other
services rely on the token used to generate payments through the payment
link for the management of that link (deletion). This should be changed
to use a key mapping to the token allowing us to remove the token from
this endpoint entirely.

Until that work is prioritised, redact the tokens from the API call
before it can reach the Toolbox service/ controller code.