PP-6257 Add starting month for all time report
Currently we add a 'Starting month' to the report manually, which is the
month the service started taking payments.

When generating the report for all time, include a 'Starting month'
column.
Don't include this when just generating a monthly report as we do not
retrieve the data from ledger for this.