BAU Only list live services by default
When viewing "Platform services", by default only show the services that
have "Go live status" of "LIVE", and do not have the "internal" or
"archived" flags set to true.

Add a link to display all services.

This should help with support when trying to track down a service you're
looking for e.g. based on an organisation name as many services have
several old or test variants of their service.