BAU: Allow gateway acounts to update Stripe payout descriptor
Provide a similar update route for support to update the payout
descriptor for a given Stripe account.

These are currently separated to allow changing only what needs
configuration (it also allows linking to separate documentation). We may
find that practically these will always be the same as the UX can be
simplified.