PP-6659 Convert generate-dev-environment to js
Javascript is easier to read and maintain than bash. Also on concourse
we have to use the Node Debian Stretch image since Cypress won't run on
Alpine. The previous `generate-dev-environment` scripts written for `sh`
used a `-d` delimiter flag for read which is not available for the
version of `sh` nor `bash` that is packaged in the Node Debian Stretch
image.