PP-7458 Improve search filters
Add filters to the service list page to filter by "Go live status",
"Internal" and "Archived". Default to displaying live, not internal, not
archived services.

Add Nunjucks macro for select input that has 3 options, a true value, a
false value and an "all" value.

Improve the filters on the gateway accounts list page to use the new
macro and patterns. Also make it so that the the CSV download for
gateway accounts obeys the filter values.