BAU: bump axios release to patch DOS
Axios released a minor `0.21.1` release to patch a DOS issue in an
underlying dependency.

`node-fetch` required by `nivo` libraries also has a low security
warning, this is something they're slowly making releases to entirely
remove the dependency however as our code doesn't actually rely on this
aspect of the charts it has been manually bumped to a non-vulnerable
version in the package lock.

The dashboard functionality has been checked and is not impacted from
first viewing.