BAU: add problematic packages with unsafe deps
React fixed at ^16 to meet the peer dependency of `@nivo` library, this
conflicts with requirements for react-spring which nivo actually uses.
This entire tangle should be moved out of this repository and into its
own component repository where it can be tracked and then resolved as
dependencies slowly updated. This shouldn't be something Toolbox devs
need to think about at the top level.

Dependencies installed with `--leagcy-peer-deps` because of tangled peer
dependencies not playing nicely with npm 7.
- react-spring
- react-inlinesvg
- @nivo/core, @nivo/line, @nice/stream

React spring issue causes a conflict with the lasest available
@types/react see https://github.com/pmndrs/react-spring/issues/1102.

Because of this it has been fixed a few versions back, when react-spring
updates to the next release candidate (4) or has a formal major version
upgrade this restriction should be lifted.