PP-7931 Generate performance page HTML
Add a route to generate the HTML for the static performance page.

This generates the main body of the page and fills in the statistics
about the services and organisations using Pay. This is returned as raw
HTML to be copied into the product page file and released.

It leaves the payment statistics blank as these currently need to be
queried directly from the database due to performance concerns.

Link to this page from a page containing instructions for updating the
product page.