PP-8319: bump chart libraries and run with --legacy-deps
Runs npm install with `--legacy-peer-deps`, the `react-spring`
dependency used by some of Nivos core libraries had a conflict that
can't be resolved with npm 7s installation. This should be resolved as
those libraries update.

Move our requirement for `react-spring` to an alias to allow it be
upgraded independantly and avoid the conflict mess.