PP-8192: restrict provider switching if already enabled
Add backend validation to top new credentials being added when the
account is already set up for switching. There is already a warning
but this should prevent any possibility of multiple non-terminal
(pre-ACTIVE) credentials existing on any given account.

Show the status of an accounts switching on the account detail page.