BAU: fix stripe account id not avaiable for stripe descriptors
The gateway account credential data model changed when credentials were
decoupled from accounts themselves. This relies on the top level
credential which is no longer avilable.

Update to use `stripe-account` endpoint to get the correct credential,
this should eventually be updated to parse all available
credentials/ integrations when that pattern is formalised.