BAU: allow users to be added to internal services
Add page to allow support and pay administrators to add users to
internal services.

It is often useful to use an existing internal service to test a
specific feature in each of the test, staging and production
environments. Access to those accounts currently involves asking someone
else to sign in and invite you through the admin tool.

Toolbox will restrict the operation to:
- services that have the `internal` flag set to true

Admin users will throw an appropriate error if:
- the service doesn't exist
- the user already has access to the service