BAU Use new ledger performance report endpoint
A new endpoint was added to get performance statistics using the ledger
`transaction_summary` table.

We had reverted to using the legacy endpoint for the performance page
JSON download, which queries the `transaction` table resulting in poor
performance and burst balance drops. This was done by mistake as the
same ledger client method is used by the live payments dashboard.

Create a separate client method that will query the legacy endpoint just
for use by the live payments dashboard.

with @SandorArpa