BAU Display outstanding Stripe setup tasks
For Stripe gateway account, display a list of the Stripe setup steps the
service is yet to complete. This will help us determine what we've
submitted to Stripe on their behalf as sometimes it can be unclear from
the Stripe dashbaord.

Separate out the table of gateway account details into smaller tables to
make it easier to visually parse:
- Gateway account details
- PSP details
- Self-serve settings
- Toolbox configurable settings