BAU Set government_entity_document flag when Stripe test account created
When we create a Stripe test account, we set all the setup flags to
completed so that we don't ask services to provide any additional
information, as all dummy information required to verify a test account
is provided upfront.

We recently added the government_entity_document flag, but weren't
setting this when creating test accounts.