PP-9587 Remove admin requirement for disabling accounts
Administative permissions should only be required in toolbox when
ability to access or edit a resource poses a high security risk.

As there is not a high security risk from being able to disable an
account, allow logged in non-admin users to perform this action.

Add a code comment on the `administrative` middleware function to
indicate when this should be used.