BAU — Put requirements for statement descriptors in more places
Several forms in toolbox ask for a bank statement descriptor for
Stripe services, either to appear on the paying user’s statement,
or the service’s statement when they receive a payout, or both.

Consistently specify the requirements for the length and allowed
characters for statement descriptors.

Source: https://stripe.com/docs/statement-descriptors#requirements