BAU: manage dashboard connection state
The live payments dashboard currently doesn't handle failures, usually
this will result in it "hanging" or no longer showing current events as
the main event loop has stopped running.

Common failure scenarios are:
- Intermittent WiFi
- Connecting to another known network not on an allowed domain list
- The tab being unfocussed for a long period of time causing the main
  event loop to not be run

Introduce a connection state, when there are any failures when
communicating with the network, reset the connection and have the
dashboard reconnect when possible. Backoff on connection attempts until
a specified maximum number of attempts.

This also allows us to appropriately reconnect when crossing a day
boundary.

The connection state is reflected visually to help anyone running the
dashboard to have some idea of what has happened without digging into
the console:
- the header highlight colour
- a pending result component

This commit cleans up some of the "event window" logic that has
particularly confusing with the current code, there are many more
simplifications that can be made but that could be done with further
refactoring.

Next steps are to:
- move all of the code for the live payments out of this repository and
  just include the built version at runtime
- when moved, include basic component and unit tests to give a bit more
  of a guarantee about health

The code can then be firmed up by:
- moving all the bundled TSX into separate and clear components
- removing any state outside of React state
- adding additional testing to guard against regressions