PP-9958 Add support for more permission levels
Previously there were only two permission levels for authenticated users
accessing secured routes: permitted and admin.

We want to introduce a third view only permission level which will allow
users to access most routes but not perform any actions that will
modify resources.

Set the permission level of a user then we authenticate a user to an
enum value - one of VIEW_ONLY, USER_SUPPORT or ADMIN.

When getting the `secured` middleware when registering the routes, pass
a PermissionLevel value. This sets the required PermissionLevel for
users trying to access this route. After checking that there is an
authenticated user, the middleware checks that the user has a permission
level equal to or greater than the required level.

Remove the `administrative` middleware and replace it by getting the
`secured` middleware with a PermissionLevel of ADMIN.

For now, users will only have a PermissionLevel of USER_SUPPORT or
ADMIN, and we will add in the ability for users to have a VIEW_ONLY
permission level in a future commit.

Use Sinon for spies in unit tests rather than chai-spies to be more
consistent with other projects.