PP-10864: Remove unused options when going live
We don't use Google analytics anymore (confirmed with our analytics people).
The analytics_id is not mandatory in the gateway accounts table in connector
too, so leaving this out won't cause an error when calling the connector api to
create a gateway account from toolbox.

We also don't use epdq and smartpay provider anymore.