PP-10864: Update go live request ticket
Update go live request ticket with current go live macro when making a service
live. Unfortunately solving the ticket using the zendesk client is troublesome
at the moment as the custom fields "[Pay] Type of user" and "[Pay] Type of user
other (enter detail)" need to be populated first. This is certainly possible
but is something we can make happen later.