BAU: Fix incorrect service live flag on go lives
After updating the service on the backend, we're logging values from a
service object fetched before the update.

Change this to log the value that we just updated the service with.

This will allow reports/ log searches to appropriately filter out
services that have been made live for internal purposes.