PP-11152 Read merchant code from new credentials structure
Worldpay credentials in the response from connector now include a
"one_off_customer_initiated" field that is always populated with the
one off credentials if they are set. The "merchant_id" and "username" at
the root of the credentials object will no longer be present for
Worldpay accounts that set/update credentials in the future.

Use the `one_off_customer_initated.merchant_code` field to display the
Worldpay merchant code.