PP-11634 Collect Notify service ID
For configuring email custom branding for a gateway account, add a
required field to set the Notify Service ID.

No changes in connector are required for this, as connector accepts
and stores any JSON for the `notify_settings` property for a gateway
account.

Add simple validation for the form to configure email branding using
the Joi library.