PP-11634- Add temporary Notify service ID checker utility
For PP-11634 some of the Notify service IDs were backfilled
incorrectly due to a copy-paste error.

Add a temporary utility route that will check the stored service ID
by making a request to the Notify API to send an email using the API
key set for the custom branding for a gateway account. Check
the service ID returned in the `template_uri` field in the response
matches the service ID we had stored in connector.

This route will be removed after it has served its purpose.