PP-11212 Improve go-live and test account journeys
Replace the 3 buttons on the services page to create live and test
gateway accounts for Stripe/Non-Stripe with just 2 buttons:
- Go live
- Add test account

When "Go live" is clicked, determine the PSP from the go-live stage
for the service. If the service hasn't completed a go-live request,
show an error.

When "Add a test account" is clicked, show a new page with radio
buttons to select the PSP for the test account (Stripe/Worldpay/Sandbox)

On the page that asks for details about the service/gateway account
before creating the gateway account in Pay:

- remove the radio button for Live/Not live, as we already know the
answer now
- remove the radio buttons for selecting the PSP, as we already know
the answer now
- prefill the description input, but still allow this to be edited
- hide the "Stripe credentials" section if we are not adding a Stripe
account.