PP-11212 Replace most of buttons on gateway account page
Replace most of the remaining buttons on the gateway account details
page.

- All the green buttons that were links to related pages have been
replaced with a list of hyperlinks.
- The "Manage API Keys" button has been replaced by listing the
number of API keys in the "Account details" summary list and adding
a "Manage" link.
- The "Switch PSP" button has been replaced with a "Manage" link in
the PSP configuration summary list.
- The Stripe statement descriptor buttons have been replaced by
fetching and displaying the statement descriptors in the "Stripe
configuration" summary table and adding "Manage" links.