PP-12549 Fix bug where & in some links encoded as &amp;amp;
If the value of a Nunjucks variable is assigned with a set
expression and that variable is later displayed in the template,
autoescaping will convert any & characters to &amp; — so the
variable value should just use & characters.

This regression was introduced by commit
e96dc1c16b79e0baa3215cfd121d8994b15f2da6.