PP-12790 Add 3D Secure information to payment page
If a payment required 3D Secure, add a ‘3DS’ tag to its page.

Add a new 3D Secure row to the processing details table, stating
whether the payment required 3D Secure, the 3D Secure version (if
known) and any information about whether an exemption was requested
and whether or not it was honoured (if available).