PP-12790 User journey duration for some failed payments
Add how long the paying user spent on the payment pages for some
failed payments (such as when the authorisation was rejected, the
user cancelled the payment or there was an error during
authorisation). In most of these cases the user will be shown an
error page and the duration will not include how long they spent
reading this page but it still gives an idea.