PP-13064 Allow toggling corporate exemptions in Worldpay accounts
With recent work[1], we have introduced the ability to toggle corporate
exemptions for Worldpay accounts in Connector via REST patch requests.

With this change, we are now introducing a new entry in the Worldpay account
configuration, showing whether corporate exemptions are enabled or disabled,
and including a link with the text 'Manage'.

Clicking on 'Manage', the user is then shown a page explaining how Worldpay
3DS Corporate Exemptions work, with radio buttons at the bottom of the page
where the user can decide to 'Turn on or off Corporate Exemptions' and
then 'Save' the change.

Further information and screenshots in Jira[2] and in the PR[3].

[1]
https://github.com/alphagov/pay-connector/pull/5513

[2]
https://payments-platform.atlassian.net/browse/PP-13064

[3]
https://github.com/alphagov/pay-toolbox/pull/1839